refactor(models): extract LCM ORM models to sibling lcm_models.py
``backend/app/models.py`` was 544 lines and over the 500-line file budget.
The four LCM retrieval tables (``LCMSummary``, ``LCMSummarySource``,
``LCMContextItem``, ``LCMEmbedding``) make the obvious split — the
governance + MCP tables already follow this pattern via
``app/governance_models.py`` and ``app/mcp_models.py``, with re-export
blocks in ``models.py`` that keep ``from app.models import LCMSummary``
working for every existing call site.

- Move the four classes (plus their ``__table_args__``) into
  ``app/lcm_models.py`` (193 lines).
- Re-export them from ``models.py`` mirroring the governance/mcp pattern
  (existing block at the bottom of the file). No other file changes —
  alembic still discovers all tables via SQLAlchemy's metadata since
  both modules share the same ``Base``.
- Remove the file-lines exemption for ``backend/app/models.py``
  (``scripts/check-file-lines.mjs``).

``models.py`` is now 410 lines, ``lcm_models.py`` is 193 lines, no
existing imports needed to change. Verified by running the LCM test
suites (``test_lcm_embeddings.py``, ``test_lcm_condensation.py``) green.
