+++ backend/app/models.py
@@ -332,154 +332,13 @@ class Workspace(Base):
     created_at: Mapped[datetime] = mapped_column(DateTime, default=_utcnow)
 
 
-class LCMSummary(Base):
-    """A single LCM summary node — either a leaf or a condensed parent."""
-
-    __tablename__ = "lcm_summaries"
-
-    id: Mapped[uuid.UUID] = mapped_column(Uuid, primary_key=True, default=uuid.uuid4)
-    conversation_id: Mapped[uuid.UUID] = mapped_column(
-        Uuid,
-        ForeignKey("conversations.id", ondelete="CASCADE"),
-        nullable=False,
-        index=True,
-    )
-    # 0 = leaf (summarises raw messages); 1+ = condensed (summarises other
-    # summaries at depth-1).  Used by the condensation pass to decide which
-    # nodes are eligible for the next level up.
-    depth: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
-    # The actual summary prose the model produced.
-    content: Mapped[str] = mapped_column(Text, nullable=False, default="")
-    # Approximate token count of ``content`` — cached so the assembly + budget
-    # math doesn't have to re-tokenise on every turn.
-    token_count: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
-    # Which model produced this summary (e.g. ``gemini-2.5-flash-preview-05-20``).
-    # Stored so future re-compaction can pick a stronger model if needed.
-    model_id: Mapped[str | None] = mapped_column(String(128), nullable=True)
-    # "normal" | "aggressive" | "fallback" — mirrors the three-level escalation
-    # used by the upstream plugin: normal prompt first, aggressive if the
-    # output is too large, deterministic truncation if both LLM passes fail.
-    summary_kind: Mapped[str] = mapped_column(String(16), nullable=False, default="normal")
-    created_at: Mapped[datetime] = mapped_column(DateTime, default=_utcnow)
-
-
-class LCMSummarySource(Base):
-    """Edge from an :class:`LCMSummary` to one of its source items.
-
-    A source is either a :class:`ChatMessage` (when the parent is a leaf
-    summary) or another :class:`LCMSummary` (when the parent is a condensed
-    summary).  ``source_kind`` discriminates between the two so a single
-    join + filter recovers either flavour.
-    """
-
-    __tablename__ = "lcm_summary_sources"
-
-    id: Mapped[uuid.UUID] = mapped_column(Uuid, primary_key=True, default=uuid.uuid4)
-    summary_id: Mapped[uuid.UUID] = mapped_column(
-        Uuid,
-        ForeignKey("lcm_summaries.id", ondelete="CASCADE"),
-        nullable=False,
-        index=True,
-    )
-    # "message" | "summary" — string discriminator so the FK can target either.
-    source_kind: Mapped[str] = mapped_column(String(16), nullable=False)
-    source_id: Mapped[uuid.UUID] = mapped_column(Uuid, nullable=False, index=True)
-    # Position in the original ordering so re-assembly is deterministic.
-    source_ordinal: Mapped[int] = mapped_column(Integer, nullable=False)
-
-
-class LCMContextItem(Base):
-    """One entry in the assembled context list for a conversation.
-
-    Walking this table in ``ordinal`` order produces the sequence of items
-    fed to the provider every turn.  Each row points at either a raw
-    :class:`ChatMessage` or a compacted :class:`LCMSummary`; the chat
-    router resolves the actual content at assembly time.
-
-    Compaction rewrites this list in place: a contiguous range of
-    ``item_kind="message"`` rows is replaced by a single
-    ``item_kind="summary"`` row, and the ordinals are renumbered so the
-    list stays dense.
-    """
-
-    __tablename__ = "lcm_context_items"
-    __table_args__ = (
-        UniqueConstraint(
-            "conversation_id",
-            "ordinal",
-            name="uq_lcm_context_items_conv_ordinal",
-        ),
-    )
-
-    id: Mapped[uuid.UUID] = mapped_column(Uuid, primary_key=True, default=uuid.uuid4)
-    conversation_id: Mapped[uuid.UUID] = mapped_column(
-        Uuid,
-        ForeignKey("conversations.id", ondelete="CASCADE"),
-        nullable=False,
-        index=True,
-    )
-    # Position in the assembled list.  Gaps are harmless; ingest_message
-    # uses max(ordinal)+1 so no dense renumbering after compaction.
-    ordinal: Mapped[int] = mapped_column(Integer, nullable=False)
-    # "message" | "summary" — the discriminator the assembly step uses to
-    # decide which lookup to perform.
-    item_kind: Mapped[str] = mapped_column(String(16), nullable=False)
-    # FK target into either chat_messages or lcm_summaries depending on
-    # ``item_kind``.  Cascades happen via the parent conversation_id FK.
-    item_id: Mapped[uuid.UUID] = mapped_column(Uuid, nullable=False, index=True)
-    created_at: Mapped[datetime] = mapped_column(DateTime, default=_utcnow)
-
-
-class LCMEmbedding(Base):
-    """Vector index entry for a single LCM-searchable item.
-
-    Issue #254 (hybrid retrieval) needs a place to store one embedding
-    per ``ChatMessage`` or ``LCMSummary`` so semantic search can run
-    alongside the lexical scorer.  We keep the vectors as JSON arrays
-    rather than introducing a pgvector dependency; that decision is
-    documented in the issue and revisitable if production workloads
-    demand it.  The ``content_hash`` column lets ingest skip
-    re-embedding unchanged content; the ``UniqueConstraint`` ensures
-    we only ever have one embedding per ``(conversation_id, item_kind,
-    item_id, embedding_model)`` tuple.
-    """
-
-    __tablename__ = "lcm_embeddings"
-    __table_args__ = (
-        UniqueConstraint(
-            "conversation_id",
-            "item_kind",
-            "item_id",
-            "embedding_model",
-            name="uq_lcm_embeddings_conv_kind_item_model",
-        ),
-    )
-
-    id: Mapped[uuid.UUID] = mapped_column(Uuid, primary_key=True, default=uuid.uuid4)
-    conversation_id: Mapped[uuid.UUID] = mapped_column(
-        Uuid,
-        ForeignKey("conversations.id", ondelete="CASCADE"),
-        nullable=False,
-        index=True,
-    )
-    # "message" | "summary" - mirrors LCMContextItem.item_kind so a
-    # search hit can be joined back to either source table.
-    item_kind: Mapped[str] = mapped_column(String(16), nullable=False)
-    item_id: Mapped[uuid.UUID] = mapped_column(Uuid, nullable=False, index=True)
-    # Identifier of the embedding provider that produced this vector.
-    # Stored so a future migration to a stronger model can re-embed
-    # in place without dropping the table.
-    embedding_model: Mapped[str] = mapped_column(String(128), nullable=False)
-    # The embedding itself - stored as a JSON array of floats.  Length
-    # is fixed per ``embedding_model`` but the database does not
-    # enforce that; the ingest pipeline checks it.
-    embedding: Mapped[list[float]] = mapped_column(JSON, nullable=False, default=list)
-    # Content hash used to skip re-embedding when the underlying text
-    # has not changed.  SHA-256 hex digest of the embedded text.
-    content_hash: Mapped[str] = mapped_column(String(64), nullable=False, default="")
-    created_at: Mapped[datetime] = mapped_column(DateTime, default=_utcnow)
-    updated_at: Mapped[datetime] = mapped_column(DateTime, default=_utcnow)
-
+# ---------------------------------------------------------------------------
+# LCM retrieval substrate (PR #258)
+#
+# Implementations live in :mod:`app.lcm_models` to keep this file under
+# the project's 500-line budget. Re-exported here so existing imports
+# (``from app.models import LCMSummary``) keep working.
+# ---------------------------------------------------------------------------
 
 # ---------------------------------------------------------------------------
 # Governance + ops platform (PRs 01-12)
@@ -488,12 +347,17 @@ class LCMEmbedding(Base):
 # under the project's 500-line budget. Re-exported here so existing
 # imports (`from app.models import AuditEvent`) keep working.
 # ---------------------------------------------------------------------------
-
 from .governance_models import (  # noqa: E402
     AuditEvent,
     CostLedger,
     ScheduledJob,
     WebhookEventRecord,
+)
+from .lcm_models import (  # noqa: E402
+    LCMContextItem,
+    LCMEmbedding,
+    LCMSummary,
+    LCMSummarySource,
 )
 from .mcp_models import McpServer  # noqa: E402  (re-export — see mcp_models.py)
